Create README.md
We are planning to develop application for Android devices which will record voice call conversations and save them on a mobile phone storage of the user. We decided to build this application because it will be useful for many reasons. Sometimes outside factors distract us and we can not understand message that is told to us. With this application you could also play some important calls whenever you want, refresh your memories with some old conversations or even you can use it as evidence. When we successfully complete this project we will share it on Google Play and wait for user reviews. If application gets good reviews we would make upgraded version with some new features for ex. Cloud support – for companies which have more frequent voice calls and record them for improving their own services. This version of application would not be free, so this is the way how we could earn some money.